Install plugins at runtime
Prior to this patch, there was an update to gitbook and where it
looks for plugins. Since the Google Analytics plugin was not installed
in the local "node_modules" folder or the global node_modules folder,
it failed to find the plugin.
It can be icky to keep trying to pre-emptively installing plugins,
instead let gitbook handle installation of plugins via it's `gitbook
install` command at runtime instead of the npm install command at image
build